flaskapp/requirements.txt
alembic==1.8.1
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.10.3
six==1.16.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
